pyproject.toml
[project]
name = "consulting-agent-with-sdk"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "claude-agent-sdk>=0.1.51",
    "langchain-openai>=1.1.13",
    "mcp[cli]>=1.26.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "tavily-python>=0.7.23",
    "langgraph>=1.0.0",
    "langchain-anthropic>=0.3.0",
    "langchain-community>=0.3.0",
    "crewai>=1.0.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.15.8",
]
